Pulito Ipotesi e aggiunto Sostituzione

+++ src/windows/AnalisiFrequenzeJFrame.java
@@ -44,7 +44,7 @@ public class AnalisiFrequenzeJFrame extends javax.swing.JFrame {
             }
         ) {
             Class[] types = new Class [] {
-                java.lang.String.class, java.lang.Integer.class, java.lang.Integer.class
+                java.lang.String.class, java.lang.Float.class, java.lang.Float.class
             };
             boolean[] canEdit = new boolean [] {
                 false, false, false
